Setup plugins for performing code coverage
This sets up the repo to perform code coverage with the `check` target on gradle. We're also upgrading to gradle to 3.4 which adds support for `jacocoTestReport` and `jacocoTestCoverageVerification` targets.

We comment out the settings.gradle to just have a focus in the exercise for the monolith. We might uncomment that again later.