We need some sub-projects to have WIP coverage testing
Not all the sub-projects are ready for full coverage testing, so we're 
reducing the testing point lower than the highest testing point for 
other projects so that they always pass in favor of projects that are 
ready. If we test below that limit, then testing coverage can apply for 
these sub projects.